Generate CA for cilium in OpenTofu
The helm_template data source for Cilium would cause Helm to regenerate
the CA on each plan, causing tofu to drift its downstream resource,
cubone's Talos machine configuration. This commit moves the CA
generation to the tofu module, so tofu will manage the PKI and remove
that constant drift.

The same was happening for Hubble's key & cert, but instead we have the
Helm chart create a CronJob for to regen the cert.